requirements.txt
"""
torch==2.0.1
torchvision==0.15.2
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
Pillow==9.3.0
"""